+++ dev/errors.R
@@ -2,6 +2,12 @@
 
 library(tabulapdf)
 
-out <- extract_tables("inst/examples/data.pdf", pages = 1, output = "data.frame")
+out <- extract_tables("inst/examples/data.pdf", pages = 1, output = "tibble")
+
+class(out)
+
+class(out[[1]])
+
+library(tibble)
 
 out